# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IAmSurfer API",
    "version": "1",
    "description": "API REST versionada (/api/v1) do app mobile IAmSurfer. Autenticação via JWT Bearer (access 15min + refresh 30d rotativo). Erros no formato { \"error\": { \"code\", \"message\", \"fields\" } }. Feeds paginados por cursor: { \"items\": [...], \"next_cursor\": ... }."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
    "/notifications/count": {
      "get": {
        "operationId": "get_api_messages_api_notifications_count",
        "tags": [
          "messages"
        ],
        "summary": "GET /notifications/count",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/notifications/read": {
      "post": {
        "operationId": "post_api_messages_api_mark_notifications_read",
        "tags": [
          "messages"
        ],
        "summary": "POST /notifications/read",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/conversations": {
      "get": {
        "operationId": "get_api_messages_api_conversations",
        "tags": [
          "messages"
        ],
        "summary": "GET /conversations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/notifications": {
      "get": {
        "operationId": "get_api_messages_api_notifications",
        "tags": [
          "messages"
        ],
        "summary": "GET /notifications",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/conversations/{username}/messages": {
      "post": {
        "operationId": "post_api_messages_api_send_message",
        "tags": [
          "messages"
        ],
        "summary": "POST /conversations/{username}/messages",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/conversations/{username}": {
      "get": {
        "operationId": "get_api_messages_api_conversation",
        "tags": [
          "messages"
        ],
        "summary": "GET /conversations/{username}",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Formato de erro único da API.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Código estável, legível por máquina."
              },
              "message": {
                "type": "string",
                "description": "Mensagem humana (pt-BR)."
              },
              "fields": {
                "type": "object",
                "description": "Erros de validação por campo.",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }
}